just move common stuff to util
