+++ test/setupTestFramework.js
@@ -1,3 +1,3 @@
 /* eslint-disable no-unused-vars */
 const fs = require('fs-extra')
-const { versionRootPath } = require('../src/common/utils')
+const { versionRootPath } = require('../src/util/utils')
